==> 1_prepare_dataset.ipynb <==
import numpy as np
import pandas as pd
import math

# ITEM_METADATA DROP

#read raw Trivago Recsys 2019 item_metadata datas
item_metadata = pd.read_csv('item_metadata.csv')

#check for "Safe Hotel" inscription in the properties of each hotel
safeHotel = item_metadata['properties'].str.contains('Safe \(Hotel\)').tolist()

#keep the ids of false ones for "Safe Hotel"
drop = set()
for i in range(len(safeHotel)):
    if safeHotel[i] == False:
        drop.add(i)

#delete the stored ids from the table and create a new csv file
item_metadata = item_metadata.drop(drop)
item_metadata.to_csv("trivago_item_metadata.csv", index = False)

#read final reduced item_metadata csv
hotel = pd.read_csv('trivago_item_metadata.csv')
print(len(hotel))

# TEST DROP

#read raw Trivago Recsys 2019 train and test datas
train = pd.read_csv('train.csv')
test = pd.read_csv('test.csv')

#keep the id of "clickout item" actions, as only the those will be used
indexTest = set()
for i in range(len(test)):
    if not test['action_type'][i] == "clickout item":
        indexTest.add(i)

#delete the stored ids from the table and create a new csv file
test = test.drop(indexTest)
test.to_csv("test_action_type_drop.csv", index = False)

#read new test data and Hotel csv
test = pd.read_csv('test_action_type_drop.csv')

#delete hotel ids not found in the trivago hotel file from the impressions column of test csv
index = []
safeHotel = set(hotel['item_id'].values)
for i in range(len(test)):
    row = test['impressions'][i].split('|')
    for j in test['impressions'][i].split('|'):
        if int(j) not in safeHotel:
            row.remove(j)
    test['impressions'][i] = "|".join(row)

#İPTAL
#keep test row with empty impressions column
empty_impression = []
for i in range(len(test['impressions'])):
    if test['impressions'][i] == "":
        empty_impression.append(i)

#delete the stored ids from the table and create a new csv file
#test = test.drop(empty_impression)

#read final reduced test file
test.to_csv("last_test.csv", index = False)
test = pd.read_csv('last_test.csv')
print(len(test))

#check if reference value of a hotel is nan or not and drop nan ones
empty_reference = []
for i in range(len(test)):
    if math.isnan(test['reference'][i]):
        empty_reference.append(i)
test = test.drop(empty_reference)

#convert reference types float to int
test['reference'] = test['reference'].astype(np.int64)

#save drop reference as csv for id changes
test.to_csv("drop_reference_test.csv", index = False)
test = pd.read_csv('drop_reference_test.csv')

#delete hotel ids not found in the trivago hotel file from the reference column of test csv
reference = []
notin_item = []
safeHotel = set(hotel['item_id'].values)
for i in range(len(test)):
    if test['reference'][i] not in safeHotel:
        notin_item.append(i)

test = test.drop(notin_item)

#save final test dataset 
test.to_csv("real_test.csv", index = False)

# TRAIN DROP

#keep the id of "clickout item" actions, as only the those will be used
indexTrain = set()
for i in range(len(train)):
    if not train['action_type'][i] == "clickout item":
        indexTrain.add(i)

#delete the stored ids from the table and create a new csv file
train = train.drop(indexTrain)
train.to_csv("train_action_type_drop.csv", index = False)

#read new train data
train = pd.read_csv('train_action_type_drop.csv')

#delete hotel ids not found in the trivago hotel file from the impressions column of train csv
index = []
safeHotel = set(hotel['item_id'].values)
for i in range(len(train)):
    row = train['impressions'][i].split('|')
    for j in train['impressions'][i].split('|'):
        if int(j) not in safeHotel:
            row.remove(j)
    train['impressions'][i] = "|".join(row)

#İPTAL
#keep train row with empty impressions column
empty_impression = []
for i in range(len(train['impressions'])):
    if train['impressions'][i] == "":
        empty_impression.append(i)

#delete the stored ids from the table and create a new csv file
#train = train.drop(empty_impression)


#read final reduced train csv
train.to_csv("train_drop.csv", index = False)
train = pd.read_csv('last_train.csv')
print(len(train))

#check if reference value of a hotel is nan or not and drop nan ones
empty_reference = []
for i in range(len(train)):
    if math.isnan(train['reference'][i]):
        empty_reference.append(i)
train = train.drop(empty_reference)

#delete hotel ids not found in the trivago hotel file from the reference column of train csv
reference = []
notin_item = []
safeHotel = set(hotel['item_id'].values)
for i in range(len(train)):
    if train['reference'][i] not in safeHotel:
        notin_item.append(i)

train = train.drop(notin_item)

#save final train dataset 
train.to_csv("real_train.csv", index = False)

# SUBMISSION DROP

#read submissions csv file
popular = pd.read_csv('submission_popular.csv')

#get all "hotel ids" from trivago_item_metadata csv
ids = hotel['item_id'].values
ids

'''
delete hotel ids not found in the trivago hotel file from the 
item_recommendations column of submission_popular csv
'''

row = set()
copy_row = set()
for i in range(len(popular)):
    row = popular['item_recommendations'][i].split(" ")
    copy_row = row.copy()
    for j in range(len(copy_row)):
        if int(copy_row[j]) not in ids:
            row.remove(copy_row[j])
    popular['item_recommendations'][i] = " ".join(row)

#create new submissions csv file
popular.to_csv("final_submission.csv", index=False)

#read final_submission file
populars = pd.read_csv('final_submission.csv')
