==> 2_modify_booking.ipynb <==
import pandas as pd
import numpy as np
import math

#read booking_otel_metadata file
booking = pd.read_csv('booking_otel_metadata.csv')

#assign empty string instead of nan cells
for j in booking.columns[8:17]:
    for i in range(len(booking)):
        if type(booking[j][i]) == float:
            booking[j][i] = ""

#concatenate hotel properties columns into a string with '|'
array = []
for i in range(len(booking)):
    line = "|".join([str(booking['service'][i]), str(booking['room'][i]),str(booking['general'][i]),str(booking['wifi'][i]),str(booking['parking'][i]),str(booking['activity'][i]),str(booking['food'][i]),str(booking['pool'][i]),str(booking['languages'][i])])
    line = line.replace('\n','|')
    array.append(line)

#assign new string properties to 'property' column
booking['property'] = array 

#removing empty properties from the property string (execute 3 times)
for i in range(len(booking)):
    if "||" in booking['property'][i]: 
        booking['property'][i] = booking['property'][i].replace("||", "|")

#delete concatenated columns from the table
del booking['room']
del booking['service']
del booking['general']
del booking['wifi']
del booking['parking']
del booking['activity']
del booking['food']
del booking['pool']
del booking['languages']

#create new booking csv file
booking.to_csv("final_booking.csv",  header=['id','name','score','city','country','url','address','img','properties'], index=False)
#read new booking file
booking = pd.read_csv("final_booking.csv")
