==> 5_prediction_with_lstm_and_knn.ipynb <==
import pickle
import numpy as np
import pandas as pd
from glob import glob
import tensorflow as tf
from tensorflow import keras
from scipy.sparse import csr_matrix
from tensorflow.keras import layers
from tensorflow.keras import Input, Model
from sklearn.neighbors import DistanceMetric
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.models import load_model
from sklearn.metrics import mean_squared_error
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import train_test_split
from tensorflow.keras.layers import Dense, LSTM, Embedding, Bidirectional

train = pd.read_csv('real_train.csv')
test = pd.read_csv('real_test.csv')

#get "impressions" and "prices" from train dataset and split them by |
train['impressions'] = train['impressions'].astype(str)
train['impressions'] = train['impressions'].map(lambda x: x.lstrip('[').rstrip(']'))
impressions = pd.DataFrame(train.impressions.str.split('|').tolist())
train.drop(columns=['impressions'], inplace = True)
train['prices'] = train['prices'].astype(str)
prices = pd.DataFrame(train.prices.str.split('|').tolist())

#concatenate impressions to train dataset as columns 
train = pd.concat([train, impressions], axis=1)
columns = train.columns.to_list()
columns[11:] = ['impressions_'+str(i) for i in range(25)]
train.columns = columns

# function that calculates weights from a neural network model
def calculate_weights(name, model):
    #calculate weights
    layer = model.get_layer(name)
    weights = layer.get_weights()[0]
    
    #normalize weight results
    weights = weights / np.linalg.norm(weights, axis = 1).reshape((-1, 1))
    
    return weights

#return hotel index 
def get_hotel_index(x):
    if x != None:
        try:
            return id_to_index[int(x)] 
        except:
            return 0
    else:
        return np.nan

embedding_model = load_model('real_embedding_50.h5')
hotel_weights = calculate_weights('hotel_embedding', embedding_model)

properties_df = pd.read_csv('real_embedding.csv')

#get item_id's and columns names from properties_df 
item_id = properties_df['item_id'].values

#create a dictionary as item_id to index
id_to_index = {item_id: idx for idx, item_id in enumerate(item_id)}
index_to_id = {idx: item_id for item_id, idx in id_to_index.items()}

# LSTM

#create a LSTM model with multi input to one output with 2 Dense Layer and Bidirectional Layer
def multi_input_one_output_LSTM(weight):
    inputs = Input(shape=(None,)) 
    
    embed = Embedding(hotel_weights.shape[0], 50, weights = [hotel_weights], trainable = False)(inputs) 
    lstm_out = Bidirectional(LSTM(25, return_sequences = False))(embed) 
    
    #outputs are set as softmax
    classifier = Dense(36, activation = 'relu')(lstm_out) 
    outputs = Dense(weight, activation = 'softmax')(classifier) 
    model = Model(inputs = [inputs], outputs = [outputs])
    
    return model

#function that prepare datasets for train_test_split()
def prepare_split_LSTM():
    #get not null rows' ids and indices
    impressionOtel_id = impressions[impressions[0] != 'nan']
    impressionOtel_index = impressionOtel_id.applymap(get_hotel_index)

    X = []
    y = []
    
    #for each line of train dataset impressions
    for i in range(impressionOtel_index.shape[0]):
        notNullCount = impressionOtel_index.iloc[i][:].count()
       
        y.append(id_to_index[train['reference'][i]])
        row = np.array(list(impressionOtel_index.iloc[i][:notNullCount].values)).reshape(-1,1)
        X.append(row)
        
    return X,y

#separates generated X and y arrays 20% as test and 80% as train
X, y = prepare_split_LSTM()
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2)

#create LSTM model and compile with sparse_categorical_crossentropy as loss and adam as optimizer 
model = multi_input_one_output_LSTM(hotel_weights.shape[0])

model.compile(loss = "sparse_categorical_crossentropy", 
              optimizer = "adam", metrics = [tf.keras.metrics.SparseTopKCategoricalAccuracy(k = 1000)])

model.summary()

X_train = [np.array(i).astype('int') for i in X_train]
y_train = [np.array([i]).astype('int') for i in y_train] 

#function that generate batches for model to training
def generator(x, y, epoch):
    for j in range(epoch):
        for i in range(len(y)):
            yield x[i].reshape(1,-1), y[i]
train_gen = generator(X_train, y_train, 5)

#fit the model with 5 epoch and with generator function
model.fit(train_gen, epochs = 5, steps_per_epoch = len(y_train))
model.save('lstm_predict.h5')

model = load_model('lstm_predict.h5')

#finding the predict result for each element of X_test and save in every 1000 to a file

maximum = []
otel_count = 1
line = 0

for i in X_test:
    temp = model.predict(i.reshape(-1,len(i)))
    temp = temp.reshape(-1)
    line += 1
    if (line % 1000) == 0:
        with open(f"/predict_results/prediction{otel_count}.pkl", 'wb') as fp:
            pickle.dump(maximum,fp)
        otel_count += 1
        maximum = []
    #append largest 20 predicted otels
    maximum.append(temp.argsort()[-20:][::-1])
    

otel_count += 1
with open(f"/predict_results/prediction{otel_count}.pkl", 'wb') as fp:
    pickle.dump(maximum,fp)

files = glob('/predict_results/*.pkl')
predict_df = pd.concat([pd.DataFrame(pd.read_pickle(fp)) for fp in files], ignore_index = True)
predict_df.to_csv('lstm_predict.csv', index = False)

lstm_predict = pd.read_csv('lstm_predict.csv')

lstm_predict

# KNN ALGORITHM

#function that prepare datasets for train_test_split()
def prepare_split_KNN():
    #get not null rows' ids and indices
    impressionOtel_id = impressions[impressions[0] != 'nan']
    impressionOtel_index = impressionOtel_id.applymap(get_hotel_index)
    
    X = []
    y = []
    
    #for each line of train dataset impressions
    for i in range(impressionOtel_index.shape[0]):
        notNullCount = impressionOtel_index.iloc[i][:].count()
        
        #checks if not null count of impressions list greater than zero 
        if(notNullCount > 0):
            y.append(id_to_index[train['reference'][i]])
            row = list(impressionOtel_index.iloc[i][:notNullCount].values)
            
            #checks if not null count of impressions list less than 25 (if true assign 0)
            if(notNullCount < 25):
                row = row + [0]*(25-notNullCount)

            X.append(np.asarray(row))
            
    return X,y

#separates generated X and y arrays 20% as test and 80% as train
X_KNN, y_KNN = prepare_split_KNN()
X_KNN_train, X_KNN_test, y_KNN_train, y_KNN_test = train_test_split(X_KNN, y_KNN, test_size = 0.2)

#calls KNeighborsClassifier function with n_neighbors parameter as 1 and fit to X_train and y_train
neigh = KNeighborsClassifier(n_neighbors= 1)
neigh.fit(X_KNN_train, y_KNN_train)

#finding the predict result for each element of X_test
predict = []

for i in X_KNN_test:
    temp = neigh.kneighbors(i.reshape(-1, len(i)))[1]
    temp = temp.reshape(-1)
    predict.append(X_KNN_train[temp[0]])

predict = pd.DataFrame(predict)
predict.to_csv("knn_predict.csv", index = False)

knn_predict = pd.read_csv("knn_predict.csv")

knn_predict

# EVALUATION

## MRR CALCULATION

#MRR navigates each row of the predictions of true_items' elements and finds id equality, 
#it takes that row position into rank.

#if it cannot find an equal id according to the metric we added, 
#when 125 (according to 20% error) and more properties find equality, 
#it takes its row position into rank.

def mrr(true_items, predictions):
  
    rank = 0.0
    count = 0
    for i in range(len(true_items)):
        predictions[i] = np.array(predictions[i])
        
        for j in range(len(predictions[i])):
 
            if(int(predictions[i][j]) == int(true_items[i])):
                count += 1
                rank += 1/(j+1)
                break
                
            else:
                result = (properties_df.loc[properties_df['item_id'] == index_to_id[int(predictions[i][j])]].values == properties_df.loc[properties_df['item_id'] == index_to_id[int(true_items[i])]].values)
                num = result.reshape(-1).tolist().count(True)
            
                if num > 125:
                    count += 1
                    rank += 1/(j+1)
                    break                 
                  
    print("count:",count)
    print("rank:",rank)
    mrr = rank / len(true_items)
    return mrr

#### MRR RESULT FOR LSTM

print(mrr(y_test, lstm_predict.values.tolist()))

#### MRR RESULT FOR KNN

print(mrr(y_KNN_test, knn_predict.values.tolist()))

## MSE CALCULATION

#similarity rate between the list of properties of 2 hotels
def properties_ratio(true_items, data, pred):
    
    final = []
    for j in range(len(true_items)):
        count = 0
        for i in range(1, len(data.columns)):
            if (list(data.loc[true_items[j]])[i:i+1][0] != list(data.loc[pred[j][0]])[i:i+1][0]): 
                count += 1

        final.append((158-count)/158)
    return final

#### MSE RESULT FOR LSTM

lstm_pred = lstm_predict['0']
lstm_pred = [float(i) for i in lstm_pred]
lstm_pred = np.array(lstm_pred).reshape(-1, 1)
lstm_test = [1.0] * len(lstm_pred['0'])

mean_squared_error(lstm_test, properties_ratio(y_test, properties_df, lstm_pred))

#### MSE RESULT FOR KNN

knn_pred = knn_predict['0']
knn_pred = [float(i) for i in knn_pred]
knn_pred = np.array(knn_pred).reshape(-1, 1)
knn_test = [1.0] * len(knn_pred['0'])

mean_squared_error(knn_test, properties_ratio(y_KNN_test, properties_df, knn_pred))
